py3/pyproject.toml
#https://grok.com/chat/8c910b33-2987-4a2d-925a-4d5e3446d68c

[build-system]
requires = ["setuptools>=40.8.0", "wheel", "anson.py3>=0.2.6"]
build-backend = "setuptools.build_meta"

[project]
name = "anclient.py3"
version = "0.1.5"
description = "Semantic-jserv Python Client with bundled io.odysz.* data"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{name = "Ody Z", email = "odys.zhou@gmail.com"}]
keywords = ["Semantics", "AnClient", "Semantier"]
requires-python = ">=3.9"
dependencies = [
    "anson.py3>=0.2.7",
    "requests",
    "typing_extensions"
]

[project.scripts]
anclient-init = "anclient.unpack_data:main"

[tool.setuptools]
packages = ["anclient"]
package-dir = { "anclient" = "src/anclient" }
package-data = { "anclient" = ["**/*.py"] }
